+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * Fixed the **generate-docs** to handle integration authentication parameter.
 * Added a validation to ensure that description and README do not contain the word 'Demisto'.
+* Fixed an issue where **generate-docs** command was displaying 'None' when credentials parameter display field configured was not configured
 
 # 1.3.9
 * Added a validation verifying that the pack's README.md file is not equal to pack description.